Remove stale config functions, combine config modules; Make-way for parsers
