+++ nucypher/config/utils.py
@@ -1,95 +1,17 @@
-import json
+import configparser
 import os
-import stat
+from typing import Tuple
 
-from .configs import NucypherConfiguration
+from web3 import IPCProvider
 
-
-def _save_private_keyfile(keypath: str, key_data: dict) -> str:
-    """
-    Creates a permissioned keyfile and save it to the local filesystem.
-    The file must be created in this call, and will fail if the path exists.
-    Returns the filepath string used to write the keyfile.
-
-    Note: getting and setting the umask is not thread-safe!
-
-    See linux open docs: http://man7.org/linux/man-pages/man2/open.2.html
-    ---------------------------------------------------------------------
-    O_CREAT - If pathname does not exist, create it as a regular file.
+from nucypher.blockchain.eth.chains import Blockchain, TesterBlockchain
+from nucypher.blockchain.eth.interfaces import EthereumContractRegistry, DeployerCircumflex, ControlCircumflex
+from nucypher.blockchain.eth.sol.compile import SolidityCompiler
+from nucypher.blockchain.eth.utilities import TemporaryEthereumContractRegistry
 
 
-    O_EXCL - Ensure that this call creates the file: if this flag is
-             specified in conjunction with O_CREAT, and pathname already
-             exists, then open() fails with the error EEXIST.
-    ---------------------------------------------------------------------
-    """
-
-    flags = os.O_WRONLY | os.O_CREAT | os.O_EXCL    # Write, Create, Non-Existing
-    mode = stat.S_IRUSR | stat.S_IWUSR              # 0o600
-
-    try:
-        keyfile_descriptor = os.open(path=keypath, flags=flags, mode=mode)
-    finally:
-        os.umask(0)  # Set the umask to 0 after opening
-
-    # Write and destroy file descriptor reference
-    with os.fdopen(keyfile_descriptor, 'wb') as keyfile:
-        keyfile.write(json.dumps(key_data))
-        output_path = keyfile.name
-
-    # TODO: output_path is an integer, who knows why?
-    del keyfile_descriptor
-    return output_path
-
-
-def _save_public_keyfile(keypath: str, key_data: bytes) -> str:
-    """
-    Creates a permissioned keyfile and save it to the local filesystem.
-    The file must be created in this call, and will fail if the path exists.
-    Returns the filepath string used to write the keyfile.
-
-    Note: getting and setting the umask is not thread-safe!
-
-    See Linux open docs: http://man7.org/linux/man-pages/man2/open.2.html
-    ---------------------------------------------------------------------
-    O_CREAT - If pathname does not exist, create it as a regular file.
-
-
-    O_EXCL - Ensure that this call creates the file: if this flag is
-             specified in conjunction with O_CREAT, and pathname already
-             exists, then open() fails with the error EEXIST.
-    ---------------------------------------------------------------------
-    """
-
-    flags = os.O_WRONLY | os.O_CREAT | os.O_EXCL    # Write, Create, Non-Existing
-    mode = stat.S_IRUSR | stat.S_IWUSR | stat.S_IRGRP | stat.S_IROTH # 0o644
-
-    try:
-        keyfile_descriptor = os.open(path=keypath, flags=flags, mode=mode)
-    finally:
-        os.umask(0) # Set the umask to 0 after opening
-
-    # Write and destroy the file descriptor reference
-    with os.fdopen(keyfile_descriptor, 'wb') as keyfile:
-        # key data should be urlsafe_base64
-        keyfile.write(key_data)
-        output_path = keyfile.name
-
-    # TODO: output_path is an integer, who knows why?
-    del keyfile_descriptor
-    return output_path
-
-
-def _parse_keyfile(keypath: str):
-    """Parses a keyfile and returns key metadata as a dict."""
-
-    with open(keypath, 'r') as keyfile:
-        try:
-            key_metadata = json.loads(keyfile)
-        except json.JSONDecodeError:
-            raise NucypherConfiguration.NucypherConfigurationError("Invalid data in keyfile {}".format(keypath))
-        else:
-            return key_metadata
+DEFAULT_CONFIG_DIR = "~"
+DEFAULT_INI_FILEPATH = './.nucypher.ini'
 
 
 def generate_confg_dir(path: str=None,) -> None:
@@ -97,8 +19,7 @@ def generate_confg_dir(path: str=None,) -> None:
     Create the configuration directory tree.
     If the directory already exists, FileExistsError is raised.
     """
-    path = path if path else NucypherConfiguration._default_configuration_directory
-
+    path = path if path else DEFAULT_CONFIG_DIR
     if not os.path.exists(path):
         os.mkdir(path, mode=0o755)
 
@@ -112,12 +33,12 @@ def validate_passphrase(passphrase) -> bool:
 
     for rule, failure_message in rules:
         if not rule:
-            raise NucypherConfiguration.NucypherConfigurationError(failure_message)
+            raise RuntimeError(failure_message)
     return True
 
 
 def check_config_tree(configuration_dir: str=None) -> bool:
-    path = configuration_dir if configuration_dir else NucypherConfiguration._default_configuration_directory
+    path = configuration_dir if configuration_dir else DEFAULT_CONFIG_DIR
     if not os.path.exists(path):
         raise FileNotFoundError('No NuCypher configuration directory found at {}.'.format(configuration_dir))
     return True
